+++ ex02/Number.hpp
@@ -6,7 +6,7 @@
 /*   By: gozon <gozon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/03 08:41:26 by gozon             #+#    #+#             */
-/*   Updated: 2025/09/04 13:24:48 by gozon            ###   ########.fr       */
+/*   Updated: 2025/09/05 08:45:39 by gozon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 class Number {
 
     private:
-        unsigned long   value;
+        int   value;
         size_t          size;
         size_t          end;
         bool*           ab;
@@ -28,7 +28,7 @@ class Number {
 
     public:
 
-        Number(unsigned long value, size_t size);
+        Number(int value, size_t size);
         Number(const Number& src);
         ~Number();
 
@@ -41,8 +41,8 @@ class Number {
         void            changeLabel(char ab, size_t index, size_t order = std::string::npos);
         size_t          getIndex(size_t order) const;
         bool            isA(size_t order) const;
-        unsigned long   getValue() const;
-        void            setValue(unsigned long value);
+        int   getValue() const;
+        void            setValue(int value);
 
         void print(size_t order) const;
